Implement tests, make report_zones function work properly
report_zones now works properly. For reference, follow 3.4.2.2 of the
NVMe ZNS command set spec, rev 1.2. The main issue was that zone_state
should be offset by 4 bits. Added a whole bunch of tests to check that
the function works properly. For some reason open_zones does not
actually mark the zone as open. I've tried this with a C version and
it doesn't work either, so it's probably an error with the arguments
being passed in (even though the return status reports no
errors). Fortunately append_zones and reset, which are the important
functions, seem to work properly, at least with a C version, so I just
need to add tests for those. Also the tests need to be ran with sudo
permissions.