Use a RWLock for zone_append
The RWLock allows multiple threads to append to the zone at the same
time if the chunk size is larger than the zone append size, and only
one thread to append if it's smaller. In addition, since the policy
doesn't check whether the zone is full before evicting it, we can
actually evict zones that are still being written to. Thus we need to
hold the lock while resetting the zone to make sure that we have
exclusive access to the zone.