Remove zone from open zones if resetting
Continuing from the previous commit, if the zone being resetted is an
open zone, this isn't actually a logic error, since we can evict zones
that are not full. In practice this is rare but still possible. This
is fine as long as we remove it and then put it back in the free zone list.