Remove stale lru elements
When evicting, we need to remove stale elements from the LRU queue. An
example:

LRU (Zone, Chunk) state:
[(1, 1), (1, 2), (2, 1), (1, 4), (3, 3)]

We choose to evict (1, 1) and (1, 2):
[(2, 1), (1, 4), (3, 3)]

(1, 4) is stale, because we reset the entire zone. This applies
whether we write in-place or not.
[(2, 1), (1, 4) { STALE }, (3, 3)]

So, we filter out all zones with 1.
[(2, 1), (3, 3)]

When we write it back to the device, we need to update the
LRU. Suppose we write it back to (2, 2). Then the LRU will look like:
[(2, 1), (3, 3), (2, 2)]