Fix endless eviction loop due to full channel
Eviction doesn't occur on Cortes when we have enough throughput, but it does work with low throughput.

We were stuck looping on cache is full

The issue comes down to eviction can't keep up with all the requests when we drain the channel. So it doesn't do enough eviction. Basically the issue occurs if you have enough throughput and we have to do multiple eviction attempts until there is enough space for all the channels and then drain.

Once I fixed that eviction does occur, however now we get a validation error invalidated read response. So I assume some data is being written back incorrectly or something. Potentially reads occurring from old data? Likely an issue with the map.