Fix race condition where reads are made from evicted locations
We had a race condition due to the inconsistency of the bucket map and underlying data

The bucket map would have had to be held during reads and writes to avoid this race condition

Implement a "pin counter", this allows us to check if the data is currently being used, and we only evict when it reaches zero.

The issue is the inconsistency between the entry and the bucket map. By doing it this way any intermediate locations will remain correct until they are done reading.

If they read while we are in process of evicting they will wait on the waiting state.